mv(exec->path): path funcs
